refactor: #172 컴포넌트 분할 (저금통 섹션, 챌린지 섹션, 차트 섹션, 지출 요약)
