+++ src/components/sections/ChallengeSection.vue
@@ -1,0 +1,142 @@
+<template>
+  <!-- 참여 중인 챌린지 섹션 -->
+  <div
+    v-if="challengeData && challengeData.status !== 'closed'"
+    class="w-[328px] h-24 rounded-2xl mt-4 flex flex-col justify-center px-6 cursor-pointer border-2 border-[#414141]"
+    @click="$emit('goToChallenge')"
+  >
+    <div class="flex items-center">
+      <div
+        class="w-12 h-12 rounded-full flex items-center justify-center mr-4"
+        :class="
+          challengeData.status === 'completed' || challengeData.status === 'failed'
+            ? ''
+            : 'bg-gray-1'
+        "
+        :style="challengeIconBackgroundStyle"
+      >
+        <FontAwesomeIcon
+          :icon="challengeIconClass"
+          class="text-xl"
+          :style="{ color: challengeIconColor }"
+        />
+      </div>
+      <div class="flex-1 flex flex-col justify-center">
+        <div class="flex items-center justify-between">
+          <div class="font-pretendard font-medium text-base text-white">
+            {{ challengeData.title }}
+          </div>
+          <div class="font-pretendard font-medium text-sm text-white ml-auto">
+            {{ challengeData.progress }}/{{ challengeData.period }}일
+          </div>
+        </div>
+        <div class="w-full bg-gray-1 rounded-full h-2 mt-1">
+          <div
+            class="h-2 rounded-full"
+            :class="challengeData.status === 'failed' ? '' : 'bg-brand'"
+            :style="progressBarStyle"
+          ></div>
+        </div>
+      </div>
+      <FontAwesomeIcon
+        :icon="faAngleRight"
+        class="text-[#c6c6c6] w-[10px] h-4 ml-2"
+      />
+    </div>
+  </div>
+
+  <!-- 참여 중인 챌린지가 없습니다 섹션 -->
+  <div
+    v-else
+    class="w-[328px] h-24 bg-[#2f2f2f] border-2 border-[#414141] rounded-2xl mt-4 flex items-center px-6 cursor-pointer"
+    @click="$emit('goToChallenge')"
+  >
+    <FontAwesomeIcon
+      :icon="fasCircleQuestion"
+      class="text-white w-12 h-12 mr-4"
+    />
+    <div class="font-pretendard font-medium text-base text-white flex-1">
+      참여 중인 챌린지가 없습니다
+    </div>
+    <FontAwesomeIcon
+      :icon="faAngleRight"
+      class="text-[#c6c6c6] w-[10px] h-4"
+    />
+  </div>
+</template>
+
+<script setup>
+import { computed } from 'vue';
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
+import {
+  faCircleQuestion as fasCircleQuestion,
+  faAngleRight,
+  faXmark,
+} from '@fortawesome/free-solid-svg-icons';
+
+const props = defineProps({
+  challengeData: {
+    type: Object,
+    default: null,
+  },
+  categoryMasterData: {
+    type: Object,
+    required: true,
+  },
+});
+
+defineEmits(['goToChallenge']);
+
+const challengeIconClass = computed(() => {
+  if (!props.challengeData) return '';
+  
+  // failed 상태일 때 xmark 아이콘 표시
+  if (props.challengeData.status === 'failed') {
+    return faXmark;
+  }
+  
+  const categoryId = props.challengeData.challenge_id;
+  const foundEntry = Object.entries(props.categoryMasterData).find(
+    ([name, cat]) => cat.id === categoryId
+  );
+
+  return foundEntry ? foundEntry[1].icon : '';
+});
+
+const challengeIconColor = computed(() => {
+  if (!props.challengeData) return '';
+  
+  // failed 상태일 때 흰색 표시
+  if (props.challengeData.status === 'failed') {
+    return '#ffffff';
+  }
+  
+  const categoryId = props.challengeData.challenge_id;
+  const foundEntry = Object.entries(props.categoryMasterData).find(
+    ([name, cat]) => cat.id === categoryId
+  );
+  return foundEntry ? foundEntry[1].color : '';
+});
+
+const challengeIconBackgroundStyle = computed(() => {
+  if (!props.challengeData) return {};
+  
+  if (props.challengeData.status === 'completed') {
+    return { backgroundColor: '#414141' };
+  } else if (props.challengeData.status === 'failed') {
+    return { backgroundColor: '#a1a1a1' };
+  }
+  return {};
+});
+
+const progressBarStyle = computed(() => {
+  if (!props.challengeData) return {};
+  
+  const progressPercentage = (props.challengeData.progress / props.challengeData.period) * 100;
+  
+  return {
+    width: `${progressPercentage}%`,
+    backgroundColor: props.challengeData.status === 'failed' ? '#858585' : '',
+  };
+});
+</script>
